sem6/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
insightface==0.7.3
onnxruntime==1.15.1
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.1.1